# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SerialCuts"
  },
  "paths": {
...
    "/api/characters/{character_id}/photos/{photo_index}": {
      "get": {
        "summary": "Character Photo",
        "operationId": "character_photo_api_characters__character_id__photos__photo_index__get",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Character Id"
            }
          },
          {
            "name": "photo_index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Photo Index"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Character Photo",
        "operationId": "delete_character_photo_api_characters__character_id__photos__photo_index__delete",
        "parameters": [
          {
            "name": "character_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Character Id"
            }
          },
          {
            "name": "photo_index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Photo Index"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CharacterRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "season_id": {
            "type": "integer",
            "title": "Season Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Aliases"
          },
          "color": {
            "type": "string",
            "title": "Color"
          },
          "photo_count": {
            "type": "integer",
            "title": "Photo Count"
          },
          "photo_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Photo Urls"
          },
          "voice_sample_count": {
            "type": "integer",
            "title": "Voice Sample Count"
          },
          "narration_voice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Narration Voice"
          },
          "narration_voice_auto": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Narration Voice Auto"
          }
        },
        "type": "object",
        "required": [
          "id",
          "season_id",
          "name",
          "description",
          "aliases",
          "color",
          "photo_count",
          "photo_urls",
          "voice_sample_count"
        ],
        "title": "CharacterRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}